phase3/pyproject.toml
[project]
name = "phase3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[tool.uv.workspace]
members = [
    "backend",
]
